requirements/base.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
colorama==0.4.3
commonmark==0.9.1
Django==3.0.8
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
future==0.18.2
gunicorn==20.0.4
jsonfield==3.1.0
kombu==4.6.11
Markdown==3.2.2
pprintpp==0.4.0
psycopg2-binary==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
rich==4.2.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.10
vine==1.3.0
